Separate EntitiyFramework conversions into partial classes
